requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.3.0.23
bootstrap4==0.1.0
celery==3.1.26.post2
celerybeat-redis==0.1.5
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
crontab==0.22.9
Django==2.2.9
django-appconf==1.0.4
django-bootstrap4==2.3.1
django-celery==3.3.1
django-compressor==2.1.1
django-constance==2.7.0
django-filter==2.4.0
django-jsonfield==1.0.1
django-model-utils==2.6.1
django-mysql==2.1.0
django-picklefield==3.0.1
django-tables2==1.2.9
django-timezone-field==4.1.1
five==0.4.0
googleads==7.0.0
httplib2==0.19.0
idna==2.7
importlib-metadata==2.1.1
kombu==3.0.37
mysqlclient==1.4.6
oauth2client==4.1.3
Pillow==8.1.0
prompt-toolkit==3.0.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==2.4.7
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.6.1
pytz==2021.1
PyYAML==5.4.1
rcssmin==1.0.6
redis==2.10.5
reportlab==3.5.59
requests==2.20.0
rjsmin==1.0.12
rsa==4.7
simplejson==3.11.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
stripe==2.55.2
suds-jurko==0.6
urllib3==1.24.3
vine==1.3.0
wcwidth==0.2.5
xmltodict==0.12.0
zipp==3.4.0
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
six
